NEMO Ertel Potential Vorticity
**In progress**

The full calculation of the Ertel potential vorticity in the NEMO (https://www.nemo-ocean.eu) simulation. Data from the calculation applied in the Western Boundary Current is shown as an example however, this computation is not confined to this region, but is applicable anywhere in the NEMO simulation. There is a manual included to explain each step of the computation to the user. The potential vorticity is calculated at 't' points on the NEMO grid (see the NEMO manual for further details with regards to the grid used).